Srch 3770 layout component (#1120)
* added components for header, footer, search bar, facets, results, identifier, and code refractor
