+++ .gitignore
@@ -49,3 +49,4 @@ mkmf.log
 /yarn-error.log
 yarn-debug.log*
 .yarn-integrity
+/node_modules/.yarn-integrity
