+++ app/javascript/components/Facets/Facets.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+interface FacetsProps {
+}
+
+export const Facets = (props: FacetsProps) => {
+  return (
+    <div id="serp-facets-wrapper">
+
+    </div>
+  );
+}
