+++ app/javascript/components/Footer/Footer.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+interface FooterProps {
+}
+
+export const Footer = (props: FooterProps) => {
+  return (
+    <div id="serp-footer-wrapper">
+
+    </div>
+  );
+}
