+++ app/javascript/components/Identifier/Identifier.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+interface IdentifierProps {
+}
+
+export const Identifier = (props: IdentifierProps) => {
+  return (
+    <div id="serp-identifier-wrapper">
+
+    </div>
+  );
+}
