+++ app/javascript/test/SearchResultsLayout.test.tsx
@@ -1,17 +1,17 @@
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 import * as React from 'react';
-import ResultsPage from '../components/ResultsPage';
+import SearchResultsLayout from '../components/SearchResultsLayout';
 
-describe('ResultsPage', () => {
+describe('SearchResultsLayout', () => {
   it('shows a message when there are no results', () => {
-    render(<ResultsPage params="foo" results={[]} vertical="web" />);
+    render(<SearchResultsLayout params="foo" results={[]} vertical="web" />);
     const message = screen.getByText(/Please enter a search term in the box above./i);
     expect(message).toBeInTheDocument();
   });
 
   it('renders all relevant links', () => {
-    render(<ResultsPage params="foo" results={[]} vertical="web" />);
+    render(<SearchResultsLayout params="foo" results={[]} vertical="web" />);
     const privacyPolicy = screen.getAllByText(/Privacy policy/i)[0];
     const updates = screen.getAllByText(/Latest updates/i)[0];
     const everything = screen.getByText(/Everything/i);
@@ -28,7 +28,7 @@ describe('ResultsPage', () => {
 
   it('renders search results', () => {
     const results = [{ 'title': 'test result 1', 'unescapedUrl': 'https://www.search.gov', 'content': 'result body' }];
-    render(<ResultsPage params="foo" results={results} vertical="web" />);
+    render(<SearchResultsLayout params="foo" results={results} vertical="web" />);
     const resultTitle = screen.getByText(/test result 1/i);
     const resultUrl = screen.getByText(/https\:\/\/www.search.gov/i);
     const resultBody = screen.getByText(/result body/i);
@@ -39,7 +39,7 @@ describe('ResultsPage', () => {
 
   it('renders image search results', () => {
     const results = [{ 'title': 'test result 1', 'thumbnail': { 'url': 'https://www.search.gov/test_image.png' } }];
-    render(<ResultsPage params="foo" results={results} vertical="image" />);
+    render(<SearchResultsLayout params="foo" results={results} vertical="image" />);
     const resultTitle = screen.getByText(/test result 1/i);
     const img = [...document.querySelectorAll("img")].pop() as HTMLImageElement;
     expect(resultTitle).toBeInTheDocument();
