+++ app/views/searches/index_redesign.html.haml
@@ -1,1 +1,1 @@
-= react_component("ResultsPage", { results: @search.results, params: @search_params, vertical: @search_vertical })
+= react_component("SearchResultsLayout", { results: @search.results, params: @search_params, vertical: @search_vertical })
